+++ RELEASE_NOTES.rst
@@ -8,6 +8,13 @@ Miro Community 1.9 Release Notes
   backwards-compatible from the backend, the user experience and
   template contexts should be the same.
 
+Miro Community 1.8.4 Release Notes
+==================================
+
+* Corrected database referencing during source imports.
+* Eliminated a thumbnail vs. source import race condition which was
+  re-marking videos as pending.
+
 Miro Community 1.8.3 Release Notes
 ==================================